\name{GWASbyChr-package}
\alias{GWASbyChr-package}
\alias{GWASbyChr}
\docType{package}
\title{Utilities for managing GWAS data in separate files for each chromosome}

\description{
Classes and methods to descrive GWAS genotype data, annotation, and association results stored in files by chromosome
}

\author{
  Stephanie Gogarten, Adrienne Stilp
  
  Maintainer: Stephanie M. Gogarten \email{sdmorris@u.washington.edu}
}

\keyword{package}
